Added gateway
